src/preprocessing/requirements.txt
asttokens==2.2.1
backcall==0.2.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
comm==0.1.4
decorator==5.1.1
dnspython==2.4.2
executing==1.2.0
filelock==3.12.2
fsspec==2023.6.0
huggingface-hub==0.16.4
idna==3.4
ipykernel==6.25.1
ipython==8.14.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.5.7
networkx==3.1
nltk==3.8.1
numpy==1.25.2
packaging==23.1
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
Pygments==2.16.1
pymongo==4.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
pyzmq==25.1.1
regex==2023.8.8
requests==2.31.0
safetensors==0.3.2
scikit-learn==1.3.0
scipy==1.11.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
stack-data==0.6.2
sympy==1.12
textblob==0.17.1
textblob-fr==0.2.0
textblob-nl==0.0.1
threadpoolctl==3.2.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.66.1
traitlets==5.9.0
transformers==4.31.0
typing_extensions==4.7.1
urllib3==2.0.4
wcwidth==0.2.6
click==8.1.6
joblib==1.3.2
nltk==3.8.1
regex==2023.8.8
textblob==0.17.1
textblob-fr==0.2.0
textblob-nl==0.0.1
